refactor(main): refactor some functions and adding an spec
skip

skip
